Move requestManager to separate module
