+++ p2p/requestManager/requestManager.go
@@ -1,4 +1,4 @@
-package protocol
+package requestManager
 
 import (
 	"crypto/rand"
